refactor: Extract alignment and character identity helpers into dndBeyond-identity.ts (closes #159)
- Create lib/import/dndBeyond-identity.ts with private constants (CANONICAL_HOST,
  CHARACTER_PATH_PATTERN, ALIGNMENT_ID_MAP), private interface (CharacterIdentity),
  private helpers (parseUrlOrThrow, isSupportedDndBeyondHostname), and exported
  functions (parseDndBeyondCharacterUrl, requireCharacterIdentity,
  buildNormalizationWarnings, normalizeAlignmentId)
- Remove all moved code from lib/dndBeyondCharacterImport.ts; add imports from new module
- Rename normalizeAlignment -> normalizeAlignmentId to avoid collision with lib/types.ts
- Re-export parseDndBeyondCharacterUrl from dndBeyondCharacterImport.ts for backward compatibility
- Use minimal structural types in new module to avoid circular imports
- All 1169 unit tests pass; tsc --noEmit clean
